fixed one problem, l2 still slightly off

+++ project/cache.h
@@ -3,6 +3,8 @@
 #include<stdlib.h>
 #include<math.h>
 #include<stdio.h>
+#include <sys/ioctl.h>
+#include <unistd.h>
 
 struct Block {
   int valid;
@@ -28,7 +30,7 @@ struct Cache *dcache;
 struct Cache *l2cache;
 
 struct Cache * initcache(int cachesize, int blocksize, int ways, int hittime, int misstime);
-int readd(struct Cache* cache,unsigned long long  t, unsigned long long index, char op);
+unsigned long long readd(struct Cache* cache,unsigned long long  t, unsigned long long index, char op);
 void printcache(struct Cache *cache);
 void printstuff(struct Cache *icache, struct Cache* dcache, struct Cache* l2cache);
